User Data can now be kept if Users leave a Guild
It's a new .env setting

+++ events/GuildMemberRemove.js
@@ -1,8 +1,10 @@
 const { CreateEventListener } = require("../EventListener.js");
 const Database = require("../Database.js");
 
+const _DELETE_USERS_ON_LEAVE = process.env["DELETE_USERS_ON_LEAVE"] !== "false";
 module.exports = CreateEventListener(
     "guildMemberRemove", async member => {
-        await Database.RemoveRows("user", { "guildId": member.guild.id, "userId": member.id });
+        if (_DELETE_USERS_ON_LEAVE)
+            await Database.RemoveRows("user", { "guildId": member.guild.id, "userId": member.id });
     }
 );
